requirements.txt
Flask==3.0.3
python-pptx==0.6.23
pandas==2.2.2
requests==2.31.0
openpyxl==3.1.2
